+++ gulpfile.js
@@ -1,3 +1,5 @@
+/* eslint-disable no-undef */
+/* eslint-disable @typescript-eslint/no-var-requires */
 const gulp = require('gulp');
 const fs = require('fs-extra');
 const path = require('path');
@@ -5,11 +7,17 @@ const chalk = require('chalk');
 const archiver = require('archiver');
 const stringify = require('json-stringify-pretty-compact');
 const typescript = require('typescript');
+const mergeStream = require('merge-stream');
+const through2 = require('through2');
 
 const ts = require('gulp-typescript');
 const less = require('gulp-less');
 const sass = require('gulp-sass');
 const git = require('gulp-git');
+const gyaml = require('gulp-yaml');
+const debug = require('gulp-debug');
+const concat = require('gulp-concat');
+const rename = require('gulp-rename');
 
 const argv = require('yargs').argv;
 
@@ -57,9 +65,9 @@ function getManifest() {
  * @returns {typescript.TransformerFactory<typescript.SourceFile>}
  */
 function createTransformer() {
-	/**
-	 * @param {typescript.Node} node
-	 */
+  /**
+   * @param {typescript.Node} node
+   */
   function shouldMutateModuleSpecifier(node) {
     if (
       !typescript.isImportDeclaration(node) &&
@@ -77,38 +85,38 @@ function createTransformer() {
     return true;
   }
 
-	/**
-	 * Transforms import/export declarations to append `.js` extension
-	 * @param {typescript.TransformationContext} context
-	 */
+  /**
+   * Transforms import/export declarations to append `.js` extension
+   * @param {typescript.TransformationContext} context
+   */
   function importTransformer(context) {
     return (node) => {
-			/**
-			 * @param {typescript.Node} node
-			 */
+      /**
+       * @param {typescript.Node} node
+       */
       function visitor(node) {
         if (shouldMutateModuleSpecifier(node)) {
           if (typescript.isImportDeclaration(node)) {
             const newModuleSpecifier = typescript.createLiteral(
-              `${node.moduleSpecifier.text}.js`
+              `${node.moduleSpecifier.text}.js`,
             );
             return typescript.updateImportDeclaration(
               node,
               node.decorators,
               node.modifiers,
               node.importClause,
-              newModuleSpecifier
+              newModuleSpecifier,
             );
           } else if (typescript.isExportDeclaration(node)) {
             const newModuleSpecifier = typescript.createLiteral(
-              `${node.moduleSpecifier.text}.js`
+              `${node.moduleSpecifier.text}.js`,
             );
             return typescript.updateExportDeclaration(
               node,
               node.decorators,
               node.modifiers,
               node.exportClause,
-              newModuleSpecifier
+              newModuleSpecifier,
             );
           }
         }
@@ -136,7 +144,20 @@ const tsConfig = ts.createProject('tsconfig.json', {
  * Build TypeScript
  */
 function buildTS() {
-  return gulp.src('src/**/*.ts').pipe(tsConfig()).pipe(gulp.dest('dist'));
+  return gulp
+    .src(['src/**/*.ts', '!src/interfaces/*.ts'])
+    .pipe(tsConfig())
+    .pipe(gulp.dest('dist'));
+}
+
+/**
+ * Build YAML to json
+ */
+function buildYaml() {
+  return gulp
+    .src(['src/**/*.yml', '!src/packs/**/*.yml'])
+    .pipe(gyaml({ space: 2, safe: true }))
+    .pipe(gulp.dest('dist'));
 }
 
 /**
@@ -157,11 +178,48 @@ function buildSASS() {
 }
 
 /**
+ * Build Compendiums
+ */
+function buildPack() {
+  const packFolders = fs.readdirSync('src/packs/').filter(function (file) {
+    return fs.statSync(path.join('src/packs', file)).isDirectory();
+  });
+  function makeid(length) {
+    var result = '';
+    var characters =
+      'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+    var charactersLength = characters.length;
+    for (var i = 0; i < length; i++) {
+      result += characters.charAt(Math.floor(Math.random() * charactersLength));
+    }
+    return result;
+  }
+
+  var packs = packFolders.map(function (folder) {
+    return gulp
+      .src(path.join('src/packs/', folder, '/**/*.yml'))
+      .pipe(
+        through2.obj((file, enc, cb) => {
+          file.contents = Buffer.concat([
+            Buffer.from(`_id: ${makeid(16)}\n`),
+            file.contents,
+          ]);
+          cb(null, file);
+        }),
+      )
+      .pipe(gyaml({ space: 0, safe: true, json: true }))
+      .pipe(concat(folder + '.json'))
+      .pipe(rename(folder + '.db'))
+      .pipe(gulp.dest('dist/packs'));
+  });
+  return mergeStream.call(null, packs);
+}
+
+/**
  * Copy static files
  */
 async function copyFiles() {
   const statics = [
-    'lang',
     'fonts',
     'assets',
     'templates',
@@ -189,9 +247,14 @@ function buildWatch() {
   gulp.watch('src/**/*.less', { ignoreInitial: false }, buildLess);
   gulp.watch('src/**/*.scss', { ignoreInitial: false }, buildSASS);
   gulp.watch(
-    ['src/fonts', 'src/lang', 'src/templates', 'src/*.json'],
+    ['src/**/*.yml', '!src/packs/**/*.yml'],
     { ignoreInitial: false },
-    copyFiles
+    buildYaml,
+  );
+  gulp.watch(
+    ['src/fonts', 'src/templates', 'src/packs', 'src/assets'],
+    { ignoreInitial: false },
+    copyFiles,
   );
 }
 
@@ -214,10 +277,13 @@ async function clean() {
       'templates',
       'assets',
       'module',
+      'interfaces',
+      'enums',
+      'packs',
       `${name}.js`,
       'module.json',
       'system.json',
-      'template.json'
+      'template.json',
     );
   }
 
@@ -269,8 +335,8 @@ async function linkUserData() {
     } else {
       throw Error(
         `Could not find ${chalk.blueBright(
-          'module.json'
-        )} or ${chalk.blueBright('system.json')}`
+          'module.json',
+        )} or ${chalk.blueBright('system.json')}`,
       );
     }
 
@@ -286,14 +352,12 @@ async function linkUserData() {
 
     if (argv.clean || argv.c) {
       console.log(
-        chalk.yellow(`Removing build in ${chalk.blueBright(linkDir)}`)
+        chalk.yellow(`Removing build in ${chalk.blueBright(linkDir)}`),
       );
 
       await fs.remove(linkDir);
     } else if (!fs.existsSync(linkDir)) {
-      console.log(
-        chalk.green(`Copying build to ${chalk.blueBright(linkDir)}`)
-      );
+      console.log(chalk.green(`Copying build to ${chalk.blueBright(linkDir)}`));
       await fs.symlink(path.resolve('./dist'), linkDir);
     }
     return Promise.resolve();
@@ -331,9 +395,7 @@ async function packageBuild() {
 
       zipFile.on('close', () => {
         console.log(chalk.green(zip.pointer() + ' total bytes'));
-        console.log(
-          chalk.green(`Zip file ${zipName} has been written`)
-        );
+        console.log(chalk.green(`Zip file ${zipName} has been written`));
         return resolve();
       });
 
@@ -372,11 +434,7 @@ function updateManifest(cb) {
   if (!manifest) cb(Error(chalk.red('Manifest JSON not found')));
   if (!rawURL || !repoURL)
     cb(
-      Error(
-        chalk.red(
-          'Repository URLs not configured in foundryconfig.json'
-        )
-      )
+      Error(chalk.red('Repository URLs not configured in foundryconfig.json')),
     );
 
   try {
@@ -402,7 +460,7 @@ function updateManifest(cb) {
             substring,
             Number(major) + 1,
             Number(minor) + 1,
-            Number(patch) + 1
+            Number(patch) + 1,
           );
           if (version === 'major') {
             return `${Number(major) + 1}.0.0`;
@@ -413,7 +471,7 @@ function updateManifest(cb) {
           } else {
             return '';
           }
-        }
+        },
       );
     }
 
@@ -424,10 +482,8 @@ function updateManifest(cb) {
     if (targetVersion === currentVersion) {
       return cb(
         Error(
-          chalk.red(
-            'Error: Target version is identical to current version.'
-          )
-        )
+          chalk.red('Error: Target version is identical to current version.'),
+        ),
       );
     }
     console.log(`Updating version number to '${targetVersion}'`);
@@ -452,7 +508,7 @@ function updateManifest(cb) {
     fs.writeFileSync(
       path.join(manifest.root, manifest.name),
       prettyProjectJson,
-      'utf8'
+      'utf8',
     );
 
     return cb();
@@ -470,7 +526,7 @@ function gitCommit() {
     git.commit(`v${getManifest().file.version}`, {
       args: '-a',
       disableAppendPaths: true,
-    })
+    }),
   );
 }
 
@@ -481,13 +537,20 @@ function gitTag() {
     `Updated to ${manifest.file.version}`,
     (err) => {
       if (err) throw err;
-    }
+    },
   );
 }
 
 const execGit = gulp.series(gitAdd, gitCommit, gitTag);
 
-const execBuild = gulp.parallel(buildTS, buildLess, buildSASS, copyFiles);
+const execBuild = gulp.parallel(
+  buildTS,
+  buildLess,
+  buildSASS,
+  buildYaml,
+  copyFiles,
+  buildPack,
+);
 
 exports.build = gulp.series(clean, execBuild);
 exports.watch = buildWatch;
@@ -500,5 +563,5 @@ exports.publish = gulp.series(
   updateManifest,
   execBuild,
   packageBuild,
-  execGit
+  execGit,
 );