+++ src/module/enums/ActorTypeEnum.ts
@@ -1,0 +1,5 @@
+export enum ActorType {
+  Character = 'character',
+  NPC = 'npc',
+  Vehicle = 'vehicle',
+}